+++ core/header_config/nav_tree.py
@@ -64,7 +64,8 @@ ATLAS_HEADER_TREE = [
             tier_two_nav_items.WHY_INVEST_IN_UK,
             tier_two_nav_items.REGIONS,
             tier_two_nav_items.SECTORS,
-            tier_two_nav_items.INVESTMENT_OPPORTUNITIES
+            tier_two_nav_items.INVESTMENT_OPPORTUNITIES,
+            tier_two_nav_items.HOW_WE_CAN_HELP_INVESTMENT_ATLAS,
         ]
     ),
     NavNode(