+++ CHANGELOG.md
@@ -3,6 +3,12 @@
 ## Pre-release
 
 ### Implemented enhancements
+
+- GP2-3233 - Atlas WHy invest in uk page and sub-pages
+- NOTICKET - d-comp version bump for cookie modal fix
+- GP2-3400 - Atlas general content page
+- GP2-3403 - Atlas opportunities listing page
+- GP2-3400: Template for generic atlas content page, which can be used for new How We Can Help page
 - GP2-3399+GP2-3401 - Nav fixups
 - NOTICKET - Fixed why invest in UK url on navigation
 - NOTICKET - Amend phrasing in Speak To Us panel